Added Android Room in build.gradle; deleted AddNoteScreen.java UI. Will add in CourseDetail.java as a field instead.
