+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/Entities/Assessment.java
@@ -1,0 +1,4 @@
+package android.jmichalek.jaymichalekwguscheduler.All.Entities;
+
+public class Assessment {
+}
