+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/Entities/Course.java
@@ -1,0 +1,4 @@
+package android.jmichalek.jaymichalekwguscheduler.All.Entities;
+
+public class Course {
+}
