+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/Entities/Term.java
@@ -1,0 +1,4 @@
+package android.jmichalek.jaymichalekwguscheduler.All.Entities;
+
+public class Term {
+}
