+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/UI/AddTermScreen.java
@@ -1,4 +1,4 @@
-package android.jmichalek.jaymichalekwguscheduler.UI;
+package android.jmichalek.jaymichalekwguscheduler.All.UI;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,12 +6,12 @@ import android.jmichalek.jaymichalekwguscheduler.R;
 import android.os.Bundle;
 import android.view.MenuItem;
 
-public class AddCourseScreen extends AppCompatActivity {
+public class AddTermScreen extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_course_screen);
+        setContentView(R.layout.activity_add_term_screen);
 
         // Allows user to switch back to previous screen & retain information.
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
@@ -28,7 +28,8 @@ public class AddCourseScreen extends AppCompatActivity {
                 return true;
         }
         return super.onOptionsItemSelected(item);
-
     }
+
+    //TODO Add action bar, back button that retains previous info, design screen
 
 }
