+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/UI/CourseDetail.java
@@ -1,17 +1,19 @@
-package android.jmichalek.jaymichalekwguscheduler.UI;
+package android.jmichalek.jaymichalekwguscheduler.All.UI;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.jmichalek.jaymichalekwguscheduler.R;
 import android.os.Bundle;
 import android.view.MenuItem;
+import android.view.View;
 
-public class AddNoteScreen extends AppCompatActivity {
+public class CourseDetail extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_note_screen);
+        setContentView(R.layout.activity_course_detail);
 
         // Allows user to switch back to previous screen & retain information.
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
@@ -28,6 +30,14 @@ public class AddNoteScreen extends AppCompatActivity {
                 return true;
         }
         return super.onOptionsItemSelected(item);
+
+    }
+
+    /* This method takes user to add assessment screen from course detail screen*/
+    public void addAssessment(View view) {
+
+        Intent intent = new Intent(CourseDetail.this, AddAssessmentScreen.class);
+        startActivity(intent);
 
     }
