+++ app/src/main/java/android/jmichalek/jaymichalekwguscheduler/All/UI/TermDetail.java
@@ -1,19 +1,17 @@
-package android.jmichalek.jaymichalekwguscheduler.UI;
+package android.jmichalek.jaymichalekwguscheduler.All.UI;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.jmichalek.jaymichalekwguscheduler.R;
 import android.os.Bundle;
 import android.view.MenuItem;
-import android.view.View;
 
-public class CourseDetail extends AppCompatActivity {
+public class TermDetail extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_course_detail);
+        setContentView(R.layout.activity_term_detail);
 
         // Allows user to switch back to previous screen & retain information.
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
@@ -30,23 +28,8 @@ public class CourseDetail extends AppCompatActivity {
                 return true;
         }
         return super.onOptionsItemSelected(item);
-
     }
 
-    /* This method takes user to add assessment screen from course detail screen*/
-    public void addAssessment(View view) {
-
-        Intent intent = new Intent(CourseDetail.this, AddAssessmentScreen.class);
-        startActivity(intent);
-
-    }
-
-    /* This method takes a user to add notes screen from course detail screen.*/
-    public void addNote(View view) {
-
-        Intent intent = new Intent(CourseDetail.this, AddNoteScreen.class);
-        startActivity(intent);
-
-    }
+    //TODO Add action bar + Term Detail screen design?
 
 }
