--- app/src/main/java/android/jmichalek/jaymichalekwguscheduler/UI/TermDetail.java
@@ -1,35 +1,0 @@
-package android.jmichalek.jaymichalekwguscheduler.UI;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.jmichalek.jaymichalekwguscheduler.R;
-import android.os.Bundle;
-import android.view.MenuItem;
-
-public class TermDetail extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_term_detail);
-
-        // Allows user to switch back to previous screen & retain information.
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-    }
-
-    /* This method enables user to switch back to previous screen.*/
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                this.finish();
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    //TODO Add action bar + Term Detail screen design?
-
-}
