+++ app/src/main/res/layout-land/activity_add_course_screen.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.AddCourseScreen">
+    tools:context=".All.UI.AddCourseScreen">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
