+++ app/src/main/res/layout-land/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.MainActivity">
+    tools:context=".All.UI.MainActivity">
 
     <TextView
         android:id="@+id/textView"
