+++ app/src/main/res/layout-land/activity_term_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.TermDetail">
+    tools:context=".All.UI.TermDetail">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
