+++ app/src/main/res/layout-land/activity_term_list.xml
@@ -5,6 +5,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.TermList">
+    tools:context=".All.UI.TermList">
 
   </androidx.constraintlayout.widget.ConstraintLayout>
