+++ app/src/main/res/layout/activity_add_term_screen.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.AddTermScreen">
+    tools:context=".All.UI.AddTermScreen">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
