+++ app/src/main/res/layout/activity_course_detail.xml
@@ -16,25 +16,7 @@
         android:onClick="addAssessment"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toEndOf="@+id/btnAddNote"
         app:srcCompat="@drawable/ic_add_foreground"
         tools:ignore="SpeakableTextPresentCheck,ImageContrastCheck,DuplicateSpeakableTextCheck" />
-
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/btnAddNote"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="232dp"
-        android:layout_marginEnd="27dp"
-        android:layout_marginBottom="32dp"
-        android:clickable="true"
-        android:onClick="addNote"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/btnAddAssessment"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@drawable/ic_addnote_foreground"
-        tools:ignore="SpeakableTextPresentCheck,ImageContrastCheck"
-        android:contentDescription="@string/todo" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
