+++ app/src/main/res/layout/activity_term_list.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.TermList">
+    tools:context=".All.UI.TermList">
 
   <com.google.android.material.floatingactionbutton.FloatingActionButton
       android:id="@+id/btn_addTerm"
@@ -13,10 +13,12 @@
       android:layout_height="wrap_content"
       android:layout_marginBottom="36dp"
       android:clickable="true"
+      android:contentDescription="TODO"
       android:onClick="addTermBtn"
       app:layout_constraintBottom_toBottomOf="parent"
       app:layout_constraintEnd_toEndOf="parent"
       app:layout_constraintHorizontal_bias="0.892"
       app:layout_constraintStart_toStartOf="parent"
-      app:srcCompat="@drawable/ic_add_foreground" />
+      app:srcCompat="@drawable/ic_add_foreground"
+      tools:ignore="SpeakableTextPresentCheck,ImageContrastCheck" />
 </androidx.constraintlayout.widget.ConstraintLayout>
